Initial Code for armada added
