+++ armada_jupyter/submissions.py
@@ -43,8 +43,10 @@ class K8sResourceOptions:
 
     def __repr__(self) -> str:
         return (
-            f"K8sResourceOptions({YMLSTR.cpu}={self.cpu}, {YMLSTR.memory}='{self.memory}', "
-            f"{YMLSTR.nvidia_gpu}={self.nvidia_gpu}, {YMLSTR.amd_gpu}={self.amd_gpu})"
+            f"K8sResourceOptions({YMLSTR.cpu}={self.cpu}, "
+            f"{YMLSTR.memory}='{self.memory}', "
+            f"{YMLSTR.nvidia_gpu}={self.nvidia_gpu}, "
+            f"{YMLSTR.amd_gpu}={self.amd_gpu})"
         )
 
 
@@ -63,7 +65,10 @@ class K8sResources:
         self.requests = limits
 
     def __repr__(self) -> str:
-        return f"K8sResources({YMLSTR.limits}={self.limits}, {YMLSTR.requests}={self.requests})"
+        return (
+            f"K8sResources({YMLSTR.limits}={self.limits}, "
+            f"{YMLSTR.requests}={self.requests})"
+        )
 
 
 class Submission:
